Refactor news endpoint controller logic to new file
